pyproject.toml
[tool.poetry]
name = "video-reverse-bot"
description = ""
version = "1.0.0"
authors = ["BlindfoldedSurgery <team@blindfolded.surgery>"]
license = "MIT"
packages = [
    { include = "app", from = "src" }
]

[[tool.poetry.source]]
name = "internal"
url = "https://pypi.blindfolded.surgery"
priority = "explicit"

[tool.poetry.dependencies]
bs-config = { version = "1.0.0", source = "internal" }
certifi = "2023.11.17"
httpx = "^0.25.1"
python = "~3.11"
python-telegram-bot = "20.6"
sentry-sdk = "^1.14.0"

[tool.poetry.group.dev.dependencies]
black = "^23.0.0"
mypy = "~1.7"
pytest = "^7.4.0"
ruff = "0.1.6"

[build-system]
requires = ["poetry-core>=1.3.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
strict = true

[[tool.mypy.overrides]]
module = "tests.*"
strict = false
check_untyped_defs = true
disallow_untyped_defs = false

[tool.ruff]
extend-select = [
    "I001",
]

[tool.ruff.isort]
known-first-party = ["app"]
